Work around overflow problem in SMALL_MULTIPLY.
Newer gcc's optimizer takes the view that the result of a
multiplication of two positive numbers can't be negative, despite of
overflow.

Issue #83

+++ c/scheme48vm-prelude.h
@@ -5,7 +5,15 @@
 #include "c-mods.h"
 #include "scheme48write-barrier.h"
 
-#define SMALL_MULTIPLY(x,y) ((x) * (y))
+/* 
+ * This is to tell the C compiler that the result may be negative even
+ * though the operands are unsigned.  It is still dubious: A cast from
+ * an unsigned to an int that involves an overflow pretty much has
+ * undefined result - the C standard does not give us two's complement
+ * arithmetic.
+ */
+
+#define SMALL_MULTIPLY(x,y) ((long)((unsigned long)(x) * (unsigned long)(y)))
 
 #define	NO_ERRORS	0		/* extension to errno.h */
